refactored view and request logic out into separate files
